Issue SC-1951 feat: Course Progress implementation
